Add note about Redux props to FAQ
The FAQ states that Reselect doesn't take additional arguments, but this doesn't cover the common scenario of those additional arguments being available as Redux props. (I needed a selector that took some additional arguments (which were available as Redux props), went straight to the FAQ (overlooking the "example" section), and saw an answer that I thought meant that I was out of luck. #162 seems to be from a user who followed a similar thought process.) This change hopefully clarifies things and adds an additional place to access this information.