Speed up selector execution
When running a selector, a copy of the parameters were cloned n^2 times which added a trivial amount
of overhead for a single call. However if a selector was called thousands of times (e.x. a large list of items)
then the trivial amount of over head added upwards of around 1 second to all selector calls regardless of
cache hits/misses.