Expose selector's result to allow adhoc caching
When a selector returns a complex value it can be difficult to get
caching right. This will expose a way for selectors to see their own
previous result which provides a way for users to do things like
return the previous result if the result of the current call will be
shallowly equal to previous result.