Implement passing memoize options directly to `createSelector`
createSelector now accepts an options object as its last argument.
For now, the only field in that object is `memoizeOptions`.  This
field represents whatever the additional parameters are for the
supplied memoization function.

If this options object is supplied, `createSelector`will try to
use the `memoizeOptions` field as the params to `memoize()`.  If
the object exists but no `memoizeOptions` field, it will fall back
to whatever memoizer options were passed as args to the original
`createSelectorCreator` call.

Further complicating things: normally, `memoizeOptions` is an array
containing all the individual args. For example, `defaultMemoize`
currently takes one options arg: the `equalityCheck` function used
for comparisons.  So, to pass that directly here, you'd do:

createSelector(
  [input1, input2],
  output,
  { memoizeOptions: [ equalityCheck ] }
)

But, it's very common to want to _only_ use the _first_ options
arg, and many libs have _only_ one arg anyway (like defaultMemoize
does right now).  So, to simplify usage, we also allow passing
that first arg directly in `memoizeOptions` without an array
around it:

createSelector(
  [input1, input2],
  output,
  { memoizeOptions: equalityCheck }
)

Internally, this is done by checking to see if `memoizeOptions` is
an array or not. If it's an array, we use it as-is. If not, we
assume it must be "the first options arg" and wrap it in an array
so it can be spread via `apply()`.

This does mean that a lib that takes an array as its first options
arg will break things here, but I can't immediately see any libs
that do that - most either take `equalityCheck` or an object.

Besides, the goal of this effort is that `defaultMemoize` will
handle enough use cases that's all you'd typically use anyway.