Rewrite selector parameter extraction for correctness
This is a massive set of changes to the extraction and interaction
of input selector parameters.  Ultimately, it boils down to:

- Correctly extract all input selector params as a nested tuple
- Transpose them so we have "all args at index 0" ... "index N"
- Intersect all the params at a given index

That should give correct behavior for determine what the final
selector function will actually accept as arguments.

The guts of the type logic is in the completely rewwritten
`MergeParameters` type, which is now heavily commented to explain
just what in the world is going on.

Since this also required bringing in a bunch of additional util
types from other libs, Stack Overflow, and the TS Github issues,
I've added comments to explain their purpose and source.