Modify test to represent the existing type inference
Previous test implies that different input state types are not supported in inference, only the state of the first selector is picked. Advanced tests for different input states in selectors therefore have to be removed since typescript v4.6 and before does not support it. Additionally, parameter is added to test, to test that the parameter is included in the output selector paraemeters.