Remove generic `F` from `CreateSelectorFunction` and `createSelectorCreator` in order to simplify the types.
I ran the all the type tests without it and it seems it can be removed without sacrificing any sort of type safety on any level.
When calling `createSelectorCreator`, it doesn't need to know the type of the function it's memoizer at some point will memoize, it can be inferred later.