Rewrite, change and rename `DropFirst` to `DropFirstParameter`
Change `DropFirst` from a type that drops the first item in an array to a type that drops the first parameter in a function, and rename it to `DropFirstParameter`.
This change was mostly due to the fact that the type was essentially a duplicate of the internal utility type `Tail`.